listener/requirements.txt
SQLAlchemy==2.0.13
python-bitcoinrpc==1.0
PyMySQL==1.1.1
pydantic==2.4.2
pydantic-settings==2.2.1